+++ sdk/python/pulumi_azure/appservice/__init__.py
@@ -6,6 +6,7 @@
 from .app_service import *
 from .active_slot import *
 from .certificate import *
+from .certificate_order import *
 from .custom_hostname_binding import *
 from .plan import *
 from .slot import *
@@ -13,4 +14,5 @@ from .source_code_token import *
 from .function_app import *
 from .get_app_service import *
 from .get_certificate import *
+from .get_certificate_order import *
 from .get_app_service_plan import *