+++ sdk/python/pulumi_azure/core/__init__.py
@@ -7,6 +7,7 @@ from .resource_group import *
 from .template_deployment import *
 from .get_client_config import *
 from .get_resource_group import *
+from .get_resources import *
 from .get_subscription import *
 from .get_subscriptions import *
 from .get_user_assigned_identity import *