+++ CHANGELOG.md
@@ -2,7 +2,7 @@ CHANGELOG
 =========
 
 ## HEAD (Unreleased)
-_(none)_
+* Regenerate SDKs based on v1.36.0 of the AzureRM Terraform Provider.
 
 ---